+++ src/flowhigh/__init__.py
@@ -1,0 +1,1 @@
+from .flowhighsr import FlowHighSR
