+++ src/flowhigh/models/attend.py
@@ -4,9 +4,11 @@ from collections import namedtuple
 
 import torch
 from torch import nn, einsum
+from torch.nn import Module
 import torch.nn.functional as F
 from einops import rearrange
 
+from .pos_emb import apply_rotary_pos_emb
 from .modules import exists, default
 
 
@@ -135,3 +137,53 @@ class Attend(nn.Module):
         out = einsum(f"b h i j, b h j d -> b h i d", attn, v)
 
         return out
+
+
+# attention
+
+class MultiheadRMSNorm(Module):
+    def __init__(self, dim, heads):
+        super().__init__()
+        self.scale = dim ** 0.5
+        self.gamma = nn.Parameter(torch.ones(heads, 1, dim))
+
+    def forward(self, x):
+        return F.normalize(x, dim = -1) * self.gamma * self.scale
+
+class Attention(Module):
+    def __init__(
+        self, dim, dim_head = 64, heads = 8, dropout = 0, flash = False, qk_norm = False, qk_norm_scale = 10 ):
+        super().__init__()
+        self.heads = heads
+        dim_inner = dim_head * heads
+
+        scale = qk_norm_scale if qk_norm else None
+
+        self.attend = Attend(dropout, flash = flash, scale = scale)
+
+        self.qk_norm = qk_norm
+
+        if qk_norm:
+            self.q_norm = MultiheadRMSNorm(dim_head, heads = heads)
+            self.k_norm = MultiheadRMSNorm(dim_head, heads = heads)
+
+        self.to_qkv = nn.Linear(dim, dim_inner * 3, bias = False)
+        self.to_out = nn.Linear(dim_inner, dim, bias = False)
+
+    def forward(self, x, mask = None, rotary_emb = None):
+        h = self.heads
+
+        q, k, v = self.to_qkv(x).chunk(3, dim = -1)
+        q, k, v = map(lambda t: rearrange(t, 'b n (h d) -> b h n d', h = h), (q, k, v))
+
+        if self.qk_norm:
+            q = self.q_norm(q)
+            k = self.k_norm(k)
+
+        if exists(rotary_emb):
+            q, k = map(lambda t: apply_rotary_pos_emb(rotary_emb, t), (q, k))
+
+        out = self.attend(q, k, v, mask = mask)
+
+        out = rearrange(out, 'b h n d -> b n (h d)')
+        return self.to_out(out)
