+++ src/flowhigh/models/common.py
@@ -1,0 +1,18 @@
+from random import random
+from einops import pack, unpack
+
+
+def divisible_by(num, den):
+    return (num % den) == 0
+
+def is_odd(n):
+    return not divisible_by(n, 2)
+
+def coin_flip():
+    return random() < 0.5
+
+def pack_one(t, pattern):
+    return pack([t], pattern)
+
+def unpack_one(t, ps, pattern):
+    return unpack(t, ps, pattern)[0]
