+++ src/flowhigh/models/pos_emb.py
@@ -1,0 +1,60 @@
+import math
+from typing import Union
+
+from beartype import beartype
+from einops import rearrange
+import torch
+from torch import nn, Tensor
+from torch.nn import Module
+from torch.cuda.amp import autocast
+from .common import divisible_by
+
+
+# sinusoidal positions
+class LearnedSinusoidalPosEmb(Module):
+    """ used by @crowsonkb """
+    def __init__(self, dim):
+        super().__init__()
+        assert divisible_by(dim, 2)
+        half_dim = dim // 2
+        self.weights = nn.Parameter(torch.randn(half_dim))
+
+    def forward(self, x):
+        x = rearrange(x, 'b -> b 1')
+        freqs = x * rearrange(self.weights, 'd -> 1 d') * 2 * math.pi
+        fouriered = torch.cat((freqs.sin(), freqs.cos()), dim = -1)
+        return fouriered
+
+
+class RotaryEmbedding(Module):
+    """
+    Rotary positional embeddings
+    Reference: https://arxiv.org/abs/2104.09864
+    """
+    def __init__(self, dim, theta = 50000):
+        super().__init__()
+        inv_freq = 1.0 / (theta ** (torch.arange(0, dim, 2).float() / dim))
+        self.register_buffer("inv_freq", inv_freq)
+
+    @property
+    def device(self):
+        return self.inv_freq.device
+
+    @autocast(enabled = False)
+    @beartype
+    def forward(self, t: Union[int, Tensor]):
+        if not torch.is_tensor(t):
+            t = torch.arange(t, device = self.device)
+
+        t = t.type_as(self.inv_freq)
+        freqs = torch.einsum('i , j -> i j', t, self.inv_freq)
+        freqs = torch.cat((freqs, freqs), dim = -1)
+        return freqs
+
+def rotate_half(x):
+    x1, x2 = x.chunk(2, dim = -1)
+    return torch.cat((-x2, x1), dim = -1)
+
+@autocast(enabled = False)
+def apply_rotary_pos_emb(pos, t):
+    return t * pos.cos() + rotate_half(t) * pos.sin()
